Move to the new API
Due to the canges in the blizzard community api, an api key is now
required, in order to authenticate agains the api. Colligate will now
ask for an api key, on start up.
Added a short delay when requesting item names. This is supposed to
reduce the risk of making too many requsts.
Changed a few things in the Downloader class. We will now use https, as
required by the api.
It is now also required to provide which locale a realm uses in
settings.yaml