Main thread rework and much more.
The main thread has been reworked. It is now better equiped to handle
exceptions in the worker threads.
The querrent downside of this is that we are no longer able to
gracefully stop Colligate by writing "stop" in the CLI.
DBmanager can now close the database connection on demand with the close
method.
I have temporarily disabled wowhead item lookup in Downloader, because I
think it is broken.
The Worker recieved a new method, used to join all threads, aka wait for
everything to finish.