Remove intermidiate json files and more
The Downloader no longer saves the json to a file. Instead it simply
returns it. We now use open-uri instead of net/http since it is a bit
simpler. Also the net/http had some problems on Ubuntu Linux.
The item lookup now happens while we wait for a realm to revcieve a new
update.