PostgreSQL to MySQL
This will be the last commit for this branch for now.
Playing with PostgreSQL have been very interesting but I have just not
been able to make it perform as well as MySQL.
One good thing that came out of this experiment was that I started using
Sequel, so now I don't have to use mysql2 gem directly. Yay!