Major speed up and date fix
Before inserting we now generate an array of hashes with the auction
data. This is then passed to Sequel with multi_insert. Inserts are now
about 82% faster Wooo!
Also fixed a minor bug when moving auctions to the log. The timestamp
used in the query was generated with strftime but used %e for day
instead of %d. This would cause the query to fail on single digit days.