Removed auction trimming before parsing
Previously we re-formatted the json into a comma seperated format in
order to read the auctions, one line at a time. This was done to
preserve memory as Colligate was originally intended to run on a
Raspberry pi.
As this is no longer in scope of the project I have removed this
functionality, and all auctions are now parsed in one go. This will use
more memory but will be more robust, as we no longer depend on some
hacky regex to reformat the auction json.